requirements.txt
azure-cognitiveservices-speech==1.38.0
openai==1.37.0
playsound==1.3.0
python-dotenv==1.0.1